move some transforms to apu_converter fix transpose filter bug fix padding value order
